@@ -14,12 +14,12 @@
         "start": "npm run build && node dist/server.js"
     },
     "dependencies": {
-        "express": "^4.21.0",
-        "express-oauth-jwt": "^2.0.2"
+        "express": "^5.1.0",
+        "jose": "^6.0.10"
     },
     "devDependencies": {
-        "@types/express": "^4.17.21",
+        "@types/express": "^5.0.1",
         "@types/node": "^20.11.30",
-        "typescript": "^5.5.4"
+        "typescript": "^5.8.3"
     }
 }
