@@ -1,6 +1,6 @@
 
 import React from 'react';
-import ReactDOM from 'react-dom';
+import {createRoot} from 'react-dom/client';
 import App from './views/app/app';
 import {AppViewModel} from './views/app/appViewModel';
 
@@ -8,7 +8,6 @@ const props = {
     viewModel: new AppViewModel(),
 };
 
-ReactDOM.render(
-    <App {...props} />,
-    document.getElementById('root'),
-);
+const container = document.getElementById('root');
+const root = createRoot(container!);
+root.render(<App {...props} />);
